models/predictor.py
import joblib
import numpy as np
import os
import json
from datetime import datetime



class AirQualityModel:
    def __init__(self):
        base_dir= os.path.dirname(os.path.abspath(__file__))
        model_dir = os.path.join(base_dir, "saved_models") 
        
        
        model_path = os.path.join(model_dir, "air_quality_model.pkl")
        scaler_path = os.path.join(model_dir, "feature_scaler.pkl")
        features_path = os.path.join(model_dir, "feature_names.txt")
        stats_path = os.path.join(model_dir, "training_stats.json")
        
        self.model = joblib.load(model_path)
        self.scaler = joblib.load(scaler_path)
        
      
        
        # Load training statistics for better confidence calculation
        try:
            with open(features_path, "r") as f:
                self.feature_names = f.read().strip().split(",")    
            with open(stats_path, "r") as f:
                self.training_stats = json.load(f)    
        except (FileNotFoundError, OSError, EOFError) as e:
            raise FileNotFoundError(f"Model files not found or corrupted in {model_dir}. Please ensure all model files are present.")
    
    def predict(self, temperature, humidity, pressure, wind_speed, hour, day_of_week):
        """Simple prediction without confidence"""
        X_new = np.array([[temperature, humidity, pressure, wind_speed, hour, day_of_week]])
        X_scaled = self.scaler.transform(X_new)
        prediction = self.model.predict(X_scaled)[0]
        return max(0, prediction)
    
    def predict_with_confidence(self, temperature, humidity, pressure, wind_speed, hour, day_of_week):
        """Optimized prediction with improved confidence calculation"""
        X_new = np.array([[temperature, humidity, pressure, wind_speed, hour, day_of_week]])
        X_scaled = self.scaler.transform(X_new)
        
        # Get predictions from all trees
        all_preds = np.array([est.predict(X_scaled)[0] for est in self.model.estimators_])
        mean_pred = all_preds.mean()
        
        # Calculate multiple confidence factors
        confidence_scores = []
        
        # 1. Tree Agreement Confidence (improved)
        std_dev = all_preds.std()
        if mean_pred > 0:
            cv = std_dev / mean_pred  # Coefficient of variation
            tree_confidence = max(0.0, 1.0 - (cv * 1.5))  # Less harsh penalty
        else:
            tree_confidence = 0.0
        confidence_scores.append(tree_confidence)
        
        # 2. Feature Range Confidence
        feature_confidence = self._calculate_feature_confidence(
            temperature, humidity, pressure, wind_speed, hour, day_of_week
        )
        confidence_scores.append(feature_confidence)
        
        # 3. Prediction Range Confidence
        range_confidence = self._calculate_prediction_range_confidence(all_preds)
        confidence_scores.append(range_confidence)
        
        # 4. Time-based Confidence
        time_confidence = self._calculate_time_confidence(hour, day_of_week)
        confidence_scores.append(time_confidence)
        
        # Weighted average of confidence scores
        weights = [0.3, 0.3, 0.2, 0.2]  # Tree agreement and feature range are most important
        combined_confidence = np.average(confidence_scores, weights=weights)
        
        # Apply scaling to achieve target 60-70%+ range
        final_confidence = self._scale_confidence(combined_confidence)
        
        return max(0, mean_pred), round(final_confidence, 3)
    
    def _calculate_feature_confidence(self, temperature, humidity, pressure, wind_speed, hour, day_of_week):
        """Calculate confidence based on how similar input features are to training data"""
        if not self.training_stats:
            return 0.65  # Default reasonable confidence
        
        features = [temperature, humidity, pressure, wind_speed, hour, day_of_week]
        feature_confidences = []
        
        for i, (feature_val, feature_name) in enumerate(zip(features, self.feature_names)):
            if feature_name in self.training_stats:
                stats = self.training_stats[feature_name]
                mean_val = stats['mean']
                std_val = stats['std']
                min_val = stats['min']
                max_val = stats['max']
                
                # Check if within reasonable range (3 standard deviations)
                if std_val > 0:
                    z_score = abs((feature_val - mean_val) / std_val)
                    if z_score <= 1.0:
                        conf = 0.9  # Within 1 std dev - high confidence
                    elif z_score <= 2.0:
                        conf = 0.8  # Within 2 std dev - good confidence
                    elif z_score <= 3.0:
                        conf = 0.6  # Within 3 std dev - moderate confidence
                    else:
                        conf = 0.4  # Outside 3 std dev - low confidence
                else:
                    conf = 0.7  # Constant feature
                
                # Extra penalty for being outside training range
                if feature_val < min_val or feature_val > max_val:
                    range_penalty = min(0.3, abs(feature_val - np.clip(feature_val, min_val, max_val)) / (max_val - min_val))
                    conf = max(0.3, conf - range_penalty)
                
                feature_confidences.append(conf)
            else:
                feature_confidences.append(0.6)  # Default if no stats
        
        return np.mean(feature_confidences)
    
    def _calculate_prediction_range_confidence(self, predictions):
        """Calculate confidence based on the spread of predictions"""
        if len(predictions) < 5:
            return 0.5
        
        # Calculate interquartile range
        q25 = np.percentile(predictions, 25)
        q75 = np.percentile(predictions, 75)
        median = np.percentile(predictions, 50)
        
        if median > 0:
            iqr_relative = (q75 - q25) / median
            # Lower relative IQR = higher confidence
            confidence = max(0.3, 1.0 - iqr_relative * 0.8)
        else:
            confidence = 0.3
        
        return confidence
    
    def _calculate_time_confidence(self, hour, day_of_week):
        """Calculate confidence based on time patterns (pollution varies by time)"""
        # Air quality patterns are often more predictable at certain times
        
        # Hour-based confidence (0-23)
        if hour in [2, 3, 4, 5]:  # Early morning - more stable
            hour_conf = 0.8
        elif hour in [7, 8, 9, 17, 18, 19]:  # Rush hours - more variable
            hour_conf = 0.6
        elif hour in [10, 11, 12, 13, 14, 15, 16]:  # Daytime - moderately stable
            hour_conf = 0.7
        else:  # Evening/night
            hour_conf = 0.7
        
        # Day of week confidence (0=Monday, 6=Sunday)
        if day_of_week in [5, 6]:  # Weekend - different patterns
            day_conf = 0.7
        else:  # Weekday - more predictable
            day_conf = 0.8
        
        return (hour_conf + day_conf) / 2
    
    def _scale_confidence(self, raw_confidence):
        """Scale confidence to achieve target range of 60-85%"""
        # Apply sigmoid-like scaling to push values into desired range
        min_conf = 0.60
        max_conf = 0.87
        
        # Sigmoid transformation centered around 0.5
        sigmoid_input = (raw_confidence - 0.5) * 4  # Stretch input
        sigmoid_output = 1 / (1 + np.exp(-sigmoid_input))
        
        # Scale to target range
        scaled = min_conf + (max_conf - min_conf) * sigmoid_output
        
        # Additional boost for very high raw confidence
        if raw_confidence > 0.8:
            scaled = min(max_conf, scaled * 1.05)
        
        return scaled
    
    def get_prediction_explanation(self, temperature, humidity, pressure, wind_speed, hour, day_of_week):
        """Get detailed explanation of the prediction"""
        prediction, confidence = self.predict_with_confidence(
            temperature, humidity, pressure, wind_speed, hour, day_of_week
        )
        
        # Categorize prediction
        if prediction <= 12:
            category = "🟢 Good"
            health_impact = "Air quality is satisfactory for most people"
        elif prediction <= 35:
            category = "🟡 Moderate"
            health_impact = "Sensitive individuals may experience minor issues"
        elif prediction <= 55:
            category = "🟠 Unhealthy for Sensitive Groups"
            health_impact = "Sensitive people should limit outdoor activities"
        elif prediction <= 150:
            category = "🔴 Unhealthy"
            health_impact = "Everyone should limit outdoor activities"
        else:
            category = "🟣 Very Unhealthy"
            health_impact = "Avoid outdoor activities"
        
        # Confidence interpretation
        if confidence >= 0.8:
            conf_text = "High confidence"
        elif confidence >= 0.7:
            conf_text = "Good confidence"
        elif confidence >= 0.6:
            conf_text = "Moderate confidence"
        else:
            conf_text = "Low confidence"
        
        return {
            'prediction': prediction,
            'confidence': confidence,
            'category': category,
            'health_impact': health_impact,
            'confidence_text': conf_text,
            'timestamp': datetime.now().isoformat()
        }