requirements.txt
streamlit==1.49.1
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.15.3
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
plotly==6.3.0
streamlit-folium==0.25.1
pydeck==0.9.1
requests==2.32.5
python-dotenv==1.1.1
